UserEntity and migration added

+++ src/Backend/TaskManager.Application/UseCases/Delete/DeleteTaskUseCase.cs
@@ -19,7 +19,7 @@ public class DeleteTaskUseCase : IDeleteTaskUseCase
         _repositoryReadOnly = repositoryReadOnly;
     }
 
-    public async Task Execute(int id)
+    public async Task Execute(long id)
     {
         var result = await _repositoryReadOnly.GetById(id);
