+++ src/Backend/TaskManager.Application/UseCases/Delete/IDeleteTaskUseCase.cs
@@ -2,5 +2,5 @@
 
 public interface IDeleteTaskUseCase
 {
-    public Task Execute(int id);
+    public Task Execute(long id);
 }
