+++ src/Backend/TaskManager.Application/UseCases/Update/IUpdateTaskUseCase.cs
@@ -4,5 +4,5 @@ namespace TaskManager.Application.UseCases.Update;
 
 public interface IUpdateTaskUseCase
 {
-    public Task Execute(int id, RequestTaskJson requestTask);
+    public Task Execute(long id, RequestTaskJson requestTask);
 }
